Add files via upload
Python Password Generator
THIS SIMPLE PROJECT WAS MADE USING PYTHON LIBRARY FUNCTIONS LIKE string & random.
string.ascii_letters

The concatenation of the ascii_lowercase and ascii_uppercase constants described below. This value is not locale-dependent.
string.ascii_lowercase

The lowercase letters abcdefghijklmnopqrstuvwxyz. This value is not locale-dependent and will not change.
string.ascii_uppercase

The uppercase letters ABCDEFGHIJKLMNOPQRSTUVWXYZ. This value is not locale-dependent and will not change.
string.digits

The string 0123456789.
string.hexdigits

The string 0123456789abcdefABCDEF.
string.octdigits The string 01234567.

string.punctuation

String of ASCII characters which are considered punctuation characters in the C locale: !"#$%&'()*+,-./:;<=>?@[\]^_{|}~
Fixed